+++ apps/benchmarks/prepare.sh
@@ -16,21 +16,28 @@ fi
 [ -z "$UB_DBA" ] && UB_DBA=sa
 [ -z "$UB_DBAPWD" ] &&  UB_DBAPWD=sa
 
+if [ -z "$UB_APP" ]; then
+  export UB_APP='tpw'
+fi
+
+if [ -z "$UB_APPDATA" ]; then
+  export UB_APPDATA='./'
+fi
 
 TESTCASE='init database'
 
 npx ubcli createStore -cfg $UB_CFG -noLogo
 if [ ! $? = 0 ]; then err; fi
 
-npx ubcli initDB -host http://localhost:$PORT -cfg $UB_CFG -dba $UB_DBA -dbaPwd $UB_DBAPWD -u admin -p admin -drop -create -cd
+npx ubcli initDB -host http://localhost:$PORT -cfg $UB_CFG -dba $UB_DBA -dbaPwd $UB_DBAPWD -p admin -drop -create -cd
 if [ ! $? = 0 ]; then err; fi
 
-TESTCASE=generateDDL
-npx ubcli generateDDL -host http://localhost:$PORT -cfg $UB_CFG -u admin -p admin -autorun
+TESTCASE='generateDDL'
+npx ubcli generateDDL -host http://localhost:$PORT -cfg $UB_CFG -autorun
 if [ ! $? = 0 ]; then err; fi
 
-TESTCASE=initialize
-npx ubcli initialize -host http://localhost:$PORT -cfg $UB_CFG -u admin -p admin
+TESTCASE='initialize'
+npx ubcli initialize -host http://localhost:$PORT -cfg $UB_CFG -u root
 if [ ! $? = 0 ]; then err; fi
 
 echo Application is ready